requirements.txt
einops==0.6.1
matplotlib==3.7.1
numpy==1.26.0
opencv-python==4.7.0.72
Pillow==9.3.0
scikit-image==0.21.0
scipy==1.10.1
tensorboard==2.12.1
timm==0.6.13
torch==2.0.0+cu118
torchvision==0.15.1+cu118
tqdm==4.65.0
